wallet: create cli
Remove commands from node cli which would otherwise require an early
return and are `unreachable`.  Logically these commands belong to
wallet, so create wallet cli, and move commands to be under wallet
instead of node.
