+++ crates/wallet/Cargo.toml
@@ -9,6 +9,10 @@ bitcoin = { version = "0.32.8", features = ["rand-std"] }
 bitcoinkernel = "0.2"
 log = "0.4"
 bdk_coin_select = "0.4.1"
+clap = { version = "4", features = ["derive"]  }
 
 [dev-dependencies]
 tempfile = "3"
+
+[[bin]]
+name = "cli"
